requirements.txt
altgraph==0.17
certifi==2020.12.5
chardet==4.0.0
docutils==0.16
future==0.18.2
idna==2.10
msedge-selenium-tools==3.141.3
pefile==2019.4.18
Pillow==8.1.0
Pygments==2.7.3
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pypiwin32==223
PyQt5==5.15.2
PyQt5-sip==12.8.1
PySide2==5.15.2
pywin32==300
pywin32-ctypes==0.2.0
requests==2.25.1
selenium==3.141.0
shiboken2==5.15.2
urllib3==1.26.2
